REFACTOR(MainActivity.kt) : 바텀 시트 뷰 이동
챗 봇 바텀 시트 뷰 이동
